backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==3.13.0
django-filter==23.2
django-rename-app==0.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
loguru==0.7.0
mccabe==0.7.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
